requirements.txt
pandas
google-generativeai
python-dotenv
tqdm
scikit-learn
joblib
numpy
qdrant-client

